Branches pruned & seeddataprovider works
Still need to register the seed provider in service container, and bind it to the contract.

Provider Caching needs a complete unit test, for storingm retrieving invalidation etc. Also it's implemented as a decorator. Nice experiment but probably should be a trait.

This commit is before I switch my workflow over to having separate online repositories for home and work: home-we-blog & work-we-blog.